Change of interface and multiple fixes
