--- modules/adas/README.md
@@ -1,2 +1,0 @@
-ADAS: Advanced Driver Assistance Systems module with Forward Collision Warning
-==============================================================================
