--- modules/adas/tools/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_subdirectory(fcw_train)
-add_subdirectory(fcw_detect)
