+++ modules/xobjdetect/CMakeLists.txt
@@ -1,2 +1,3 @@
 set(the_description "Object detection algorithms")
-ocv_define_module(xobjdetect opencv_core opencv_imgproc opencv_highgui WRAP python)
+ocv_define_module(xobjdetect opencv_core opencv_imgproc opencv_highgui opencv_objdetect WRAP python)
+add_subdirectory(tools)
