+++ modules/xobjdetect/src/precomp.hpp
@@ -1,7 +1,6 @@
 /*
-
-By downloading, copying, installing or using the software you agree to this
-license. If you do not agree to this license, do not download, install,
+By downloading, copying, installing or using the software you agree to this license.
+If you do not agree to this license, do not download, install,
 copy or use the software.
 
 
@@ -9,7 +8,12 @@ copy or use the software.
                For Open Source Computer Vision Library
                        (3-clause BSD License)
 
-Copyright (C) 2013, OpenCV Foundation, all rights reserved.
+Copyright (C) 2000-2015, Intel Corporation, all rights reserved.
+Copyright (C) 2009-2011, Willow Garage Inc., all rights reserved.
+Copyright (C) 2009-2015, NVIDIA Corporation, all rights reserved.
+Copyright (C) 2010-2013, Advanced Micro Devices, Inc., all rights reserved.
+Copyright (C) 2015, OpenCV Foundation, all rights reserved.
+Copyright (C) 2015, Itseez Inc., all rights reserved.
 Third party copyrights are property of their respective owners.
 
 Redistribution and use in source and binary forms, with or without modification,
@@ -28,31 +32,48 @@ are permitted provided that the following conditions are met:
 
 This software is provided by the copyright holders and contributors "as is" and
 any express or implied warranties, including, but not limited to, the implied
-warranties of merchantability and fitness for a particular purpose are
-disclaimed. In no event shall copyright holders or contributors be liable for
-any direct, indirect, incidental, special, exemplary, or consequential damages
+warranties of merchantability and fitness for a particular purpose are disclaimed.
+In no event shall copyright holders or contributors be liable for any direct,
+indirect, incidental, special, exemplary, or consequential damages
 (including, but not limited to, procurement of substitute goods or services;
 loss of use, data, or profits; or business interruption) however caused
 and on any theory of liability, whether in contract, strict liability,
 or tort (including negligence or otherwise) arising in any way out of
 the use of this software, even if advised of the possibility of such damage.
-
 */
 
 #ifndef __OPENCV_XOBJDETECT_PRECOMP_HPP__
 #define __OPENCV_XOBJDETECT_PRECOMP_HPP__
 
 #include <opencv2/xobjdetect.hpp>
-#include <opencv2/xobjdetect/private.hpp>
 
 #include <opencv2/core/utility.hpp>
 
 #include <opencv2/imgproc.hpp>
 #include <opencv2/imgproc/types_c.h>
 
+#include <opencv2/core.hpp>
+#include <opencv2/highgui.hpp>
+
 #include <opencv2/imgcodecs.hpp>
+#include <opencv2/imgcodecs/imgcodecs_c.h>
+
+#include <opencv2/objdetect.hpp>
 
 #include <algorithm>
+#include <vector>
+#include <string>
 #include <cmath>
+#include <iostream>
+#include <fstream>
+#include <iomanip>
+#include <sstream>
+#include <cassert>
+#include <cstdio>
+
+#include "cascadeclassifier.h"
+#include "traincascade_features.h"
+#include "lbpfeatures.h"
+#include "waldboost.hpp"
 
 #endif /* __OPENCV_XOBJDETECT_PRECOMP_HPP__ */
