+++ modules/xobjdetect/src/waldboost.hpp
@@ -42,56 +42,47 @@ or tort (including negligence or otherwise) arising in any way out of
 the use of this software, even if advised of the possibility of such damage.
 */
 
-#ifndef _OPENCV_CASCADECLASSIFIER_H_
-#define _OPENCV_CASCADECLASSIFIER_H_
+#ifndef __OPENCV_XOBJDETECT_WALDBOOST_HPP__
+#define __OPENCV_XOBJDETECT_WALDBOOST_HPP__
 
-#include <ctime>
-#include "traincascade_features.h"
-#include "lbpfeatures.h"
+#include "precomp.hpp"
 
-#define CC_CASCADE_FILENAME "cascade.xml"
-#define CC_PARAMS_FILENAME "params.xml"
+namespace cv {
 
-#define CC_CASCADE_PARAMS "cascadeParams"
-#define CC_STAGE_TYPE "stageType"
-#define CC_FEATURE_TYPE "featureType"
-#define CC_HEIGHT "height"
-#define CC_WIDTH  "width"
+class WaldBoost {
+public:
+    WaldBoost(int weak_count);
+    WaldBoost();
+    std::vector<int> get_feature_indices();
 
-#define CC_STAGE_NUM    "stageNum"
-#define CC_STAGES       "stages"
-#define CC_STAGE_PARAMS "stageParams"
+    void detect(Ptr<CvFeatureEvaluator> eval,
+                const Mat& img,
+                const std::vector<float>& scales,
+                std::vector<Rect>& bboxes,
+                Mat1f& confidences);
 
-#define CC_BOOST            "BOOST"
-#define CC_BOOST_TYPE       "boostType"
-#define CC_DISCRETE_BOOST   "DAB"
-#define CC_REAL_BOOST       "RAB"
-#define CC_LOGIT_BOOST      "LB"
-#define CC_GENTLE_BOOST     "GAB"
-#define CC_MINHITRATE       "minHitRate"
-#define CC_MAXFALSEALARM    "maxFalseAlarm"
-#define CC_TRIM_RATE        "weightTrimRate"
-#define CC_MAX_DEPTH        "maxDepth"
-#define CC_WEAK_COUNT       "maxWeakCount"
-#define CC_STAGE_THRESHOLD  "stageThreshold"
-#define CC_WEAK_CLASSIFIERS "weakClassifiers"
-#define CC_INTERNAL_NODES   "internalNodes"
-#define CC_LEAF_VALUES      "leafValues"
+    void detect(Ptr<CvFeatureEvaluator> eval,
+                const Mat& img,
+                const std::vector<float>& scales,
+                std::vector<Rect>& bboxes,
+                std::vector<double>& confidences);
 
-#define CC_FEATURES       FEATURES
-#define CC_FEATURE_PARAMS "featureParams"
-#define CC_MAX_CAT_COUNT  "maxCatCount"
-#define CC_FEATURE_SIZE   "featSize"
+    void fit(Mat& data_pos, Mat& data_neg);
+    int predict(Ptr<CvFeatureEvaluator> eval, float *h) const;
+    void save(const std::string& filename);
+    void load(const std::string& filename);
+    void reset(int weak_count);
+    ~WaldBoost();
 
-#define CC_HAAR        "HAAR"
-#define CC_MODE        "mode"
-#define CC_MODE_BASIC  "BASIC"
-#define CC_MODE_CORE   "CORE"
-#define CC_MODE_ALL    "ALL"
-#define CC_RECTS       "rects"
-#define CC_TILTED      "tilted"
+private:
+    int weak_count_;
+    std::vector<float> thresholds_;
+    std::vector<float> alphas_;
+    std::vector<int> feature_indices_;
+    std::vector<int> polarities_;
+    std::vector<float> cascade_thresholds_;
+};
 
-#define CC_LBP  "LBP"
-#define CC_RECT "rect"
+}
 
 #endif
